==> Data Analysis Nanodegree/Lesson 5/1 assessing.ipynb <==
# Assessing
Use the space below to explore `all_alpha_08.csv` and `all_alpha_18.csv` to answer the quiz questions below.

import pandas as pd
import numpy as np
%matplotlib inline

df_08 = pd.read_csv("all_alpha_08.csv")
df_18 = pd.read_csv("all_alpha_18.csv")

df_08.tail()

df_18.head()

df_08.shape

df_18.shape

sum(df_08.duplicated())

sum(df_18.duplicated())

df_08.dtypes

df_18.dtypes

# Count distict values, use nunique:
df_08.nunique()

df_18.nunique()

# Count only non-null values, use count:
# rows with missing data
2404 - min(df_08.count())

1611 - min(df_18.count())

# Count total values including null values, use size attribute:
df_08.size

df_08.groupby('Fuel')['Fuel'].count()

df_18.groupby('Fuel')['Fuel'].count()
