==> Data Analysis Nanodegree/Lesson 5/4 inspect_datatypes.ipynb <==
# Inspecting Data Types
Use the space below to explore `data_08_v2.csv` and `data_18_v2.csv` to answer the quiz questions below regarding datatypes. You should've created these data files in the previous section: *Filter, Drop Nulls, Dedupe*.

import pandas as pd
df_08 = pd.read_csv("data_08_v2.csv")
df_18 = pd.read_csv("data_18_v2.csv")


df_08.head()

df_18.head()

df_08.dtypes

df_18.dtypes

df_08
