==> Data Analysis Nanodegree/Lesson 5/8 drawing_conclusions.ipynb <==
# Drawing Conclusions
Use the space below to address questions on datasets `clean_08.csv` and `clean_18.csv`. You should've created these data files in the previous section: *Fixing Data Types Pt 3*.

# load datasets
import pandas as pd
df_08 = pd.read_csv("clean_08.csv")
df_18 = pd.read_csv("clean_18.csv")

### Q1: Are more unique models using alternative sources of fuel? By how much?

df_08.shape

df_08.head()

df_08['fuel'].value_counts()

df_18.shape

df_18['fuel'].value_counts()

print(3*100/987)
print((26+26+19+12))
print((26+26+19+12)*100/832)

More unique models are using alternative sources of fuel now
In 2008 it was 3 models, now it's 83
That's an increase from 0.3% of all models to 9.98%

### Q2: How much have vehicle classes improved in fuel economy?  

df_08_cmb = df_08.groupby('veh_class')['cmb_mpg'].mean()

df_18_cmb = df_18.groupby('veh_class')['cmb_mpg'].mean()

df_08_cmb

df_18_cmb

(df_18_cmb-df_08_cmb)/df_08_cmb

### Q3: What are the characteristics of SmartWay vehicles? Have they changed over time?

df_08.groupby('smartway').mean()

df_18.groupby('smartway').mean()

### Q4: What features are associated with better fuel economy?

df_08['cmb_mpg'].mean()

df_08['good_fuel_econ'] = df_08['cmb_mpg'] > df_08['cmb_mpg'].mean()

df_08.groupby('good_fuel_econ').mean()
